LayoutTests/fast/dom/shadow/touch-event-retargeting-expected.txt

Tests to ensure that event dispatching behaves as the Shadow DOM spec describes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



  touchstart
     @div1 (target: div1) (touches: div1, div2, div3) (targetTouches: div1) (changedTouches: div1, div2, div3)
     @div2 (target: div2) (touches: div1, div2, div3) (targetTouches: div2) (changedTouches: div1, div2, div3)
     @div2-shadow-root-child (target: div2-shadow-root-child) (touches: div1, div2-shadow-root-child, div3) (targetTouches: div2-shadow-root-child) (changedTouches: div1, div2-shadow-root-child, div3)
     @div3 (target: div3) (touches: div1, div2, div3) (targetTouches: div3) (changedTouches: div1, div2, div3)
     @div3-shadow-root-child (target: div3-shadow-root-child) (touches: div1, div2, div3-shadow-root-child) (targetTouches: div3-shadow-root-child) (changedTouches: div1, div2, div3-shadow-root-child)
PASS successfullyParsed is true

TEST COMPLETE
Touch event retargeting.

foo
